working on term of repayment result
