+++ application/templates/debt_form.html
@@ -3,6 +3,7 @@
 {% block main_content %}
 
 <div class="main">
+	<h1>Add Debt</h1>
 	<p>Enter details for 1 debt below:</p>
 	<fieldset>
 		<legend>
@@ -41,9 +42,9 @@
 
 	<br>
 <!-- See example debts -->
-	<div>
-    	<button id="example_debt_summary" class="btn"><a href="./debt_summary">See Example Debt Data</a></button>
-  	</div>
+	<!-- <div>
+    	<button id="example_debt_summary" class="btn"><a href="./debt_summary">Use Example Debt Data</a></button>
+  	</div> -->
 
 </div>
