+++ application/templates/debt_summary.html
@@ -5,13 +5,24 @@
 <div class="main">
 	<div>
 		<h1>Debt Summary</h1>
+		<img src="{{ url_for('static', filename='images/anxious_woman.jpeg') }}" class="img-responsive margin" alt="woman solving debt problems" width="150px">
 
-		<p>
+		<table class="summary">
+			<tr>
+				<th>Debt Name:</th>
+				<th>&pound;</th>
+				<th>Min Repayment</th>
+				<th>% APR</th>
+			</tr>
 			{% for dictionary in debts %}
-			Debt Name: {{ dictionary['identifier']}}<br>
-			&pound;{{ '%.2f' % dictionary['amount'] }}<br>with {{ dictionary['interest'] }}%APR and repayments of &pound;{{ dictionary['repayment'] }}<br>
+			<tr>
+				<td>{{ dictionary['identifier']}}</td>
+				<td>&pound;{{ dictionary['amount'] }}.</td>
+				<td>&pound;{{ dictionary['repayment'] }}.</td>
+				<td>{{ dictionary['interest'] }}%</td>
+			</tr>
 			{% endfor %}
-		</p>
+		</table>
 	</div>
 
 	<br>
@@ -23,9 +34,8 @@
 	<br>
 
 	<div>
-		<p>Current Minimum Repayments Total &pound;{{ total }}<br>
-		Enter a monthly total that you can commit to repaying.<br>
-		This number MUST be equal to or bigger than the total minimum repayment of &pound;{{ total }}.</p>
+		<p>Current Minimum Repayment:<br>Total &pound;{{ total }}<br>
+		Enter a monthly total that you can commit to repaying.</p>
 		
 		<fieldset>
 			<legend>
