+++ application/templates/index.html
@@ -13,7 +13,7 @@
   </div>
 
   <div>
-    <table id="table">
+    <table class="intro">
       <tbody>
         <tr>
           <td><img src="{{ url_for('static', filename='images/box_stacksmall.jpg') }}"></td>
