+++ application/templates/results.html
@@ -4,37 +4,30 @@
 
 <div class="main">
     <div>
-        <h1>Results</h1>
-        <img src="{{ url_for('static', filename='images/anxious_woman.jpeg') }}" class="img-responsive margin" alt="woman solving debt problems" width="200px">
+        <h1>Results - &pound;{{ repay }}pcm</h1>
+        
     </div>
 
     <div>
-        <h3>Debt Summary</h3>
-
+        <p><img src="{{ url_for('static', filename='images/box_stacksmall.jpg') }}"> <span>STACK Approach</span></p>
+        <p>Debt free in {{ stack }}</p>
         {% for debt in debt_list %}
-        <p>Debt Id: {{ debt['identifier'] }}</p>
-        <p>Debt Amount: &pound;{{ '%.2f' % debt['amount'] }}</p>
-        <p>Minimum Monthly Repayment: {{ '%.2f' % debt['repayment'] }}</p>
+        <p>{{ debt['identifier'] }} will be paid in {{ debt['stack_years'] }} </p>
+        {% endfor %}
+
+        <p><img src="{{ url_for('static', filename='images/snowball_small.jpg') }}"> <span>SNOWBALL Approach</span></p>
+        <p>Debt free in {{ snowball }}</p>
+        {% for debt in debt_list %}
+        <p>{{ debt['identifier'] }} will be paid in {{ debt['snowball_years'] }}</p>
+        {% endfor %}
+
+        <p><img src="{{ url_for('static', filename='images/avalanche_cartoon_small.jpg') }}"> <span>AVALANCHE Approach</span></p>
+        <p>Debt free in {{ avalanche }}</p>
+        {% for debt in debt_list %}
+        <p>{{ debt['identifier'] }} will be paid in {{ debt['avalanche_years'] }}</p>
+
         {% endfor %}
     </div>
-
-    <h4>Which debt repayment approach<br>is better for you?</h4>
-
-    <p>Stack Approach</p>
-    {% for debt in debt_list %}
-    <p>{{ debt['identifier'] }} will be paid in {{ debt['stack_years'] }} </p>
-    {% endfor %}
-
-    <p>Snowball Approach</p>
-    {% for debt in debt_list %}
-    <p>{{ debt['identifier'] }} will be paid in {{ debt['snowball_years'] }}</p>
-    {% endfor %}
-
-    <p>Avalanche Approach</p>
-    {% for debt in debt_list %}
-    <p>{{ debt['identifier'] }} will be paid in {{ debt['avalanche_years'] }}</p>
-
-    {% endfor %}
 
 </div>
