+++ application/v1/debt.py
@@ -1,0 +1,12 @@
+class Debt:
+    
+    def __init__(self, identifier, amount, interest, repayment):
+        self.identifier = identifier
+        self._amount = amount
+        self._interest = interest
+        self._repayment = repayment
+    
+
+
+
+
